+++ library/src/test/java/com/evernote/android/job/BaseJobManagerTest.java
@@ -9,7 +9,7 @@ import android.support.annotation.NonNull;
 import android.test.mock.MockContext;
 
 import com.evernote.android.job.test.DummyJobs;
-import com.evernote.android.job.test.TestCat;
+import com.evernote.android.job.test.TestLogger;
 
 import org.junit.Rule;
 import org.robolectric.RuntimeEnvironment;
@@ -78,7 +78,7 @@ public abstract class BaseJobManagerTest {
     }
 
     protected Future<Job.Result> executeJobAsync(int jobId, @NonNull final Job.Result expected) throws InterruptedException {
-        final JobProxy.Common common = new JobProxy.Common(context(), TestCat.INSTANCE, jobId);
+        final JobProxy.Common common = new JobProxy.Common(context(), TestLogger.INSTANCE, jobId);
 
         final JobRequest pendingRequest = common.getPendingRequest(true, true);
         assertThat(pendingRequest).isNotNull();