+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add new `METERED` network type
 * Add new requirements battery not low and storage not low
 * Add helper job class `DailyJob` to make it easier to run jobs once a day, see #223
+* Add option in `JobConfig` to add a logger
 
 ## 1.1.12